+++ packages/migrations/src/MigrationsService.ts
@@ -61,7 +61,9 @@ export class MigrationsService {
       try {
         await func(context)
       } catch (error) {
-        logger.error('[orionjs/migrations] Error running migration, will abort transaction', error)
+        logger.error('[orionjs/migrations] Error running migration, will abort transaction', {
+          error,
+        })
         throw error
       }
     })