+++ packages/dogs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @orion-js/dogs
 
+## 4.1.9
+
+### Patch Changes
+
+- fixes types in job schedule params
+- 2ed96d2: better logger instructions
+
 ## 4.1.8
 
 ### Patch Changes